test: add margin tests

+++ tests/utils/PoRBaseTest.sol
@@ -156,7 +156,7 @@ abstract contract PoRBaseTest is Test {
 
   function _mintBacked(MockERC20 asset, uint256 amount) internal {
     _mintUnbacked(asset, amount);
-    _setPoRAnswer(asset, asset.totalSupply());
+    _setPoRAnswer(asset, int256(asset.totalSupply()));
   }
 
   function _mintUnbacked(MockERC20 asset, uint256 amount) internal {
@@ -167,13 +167,13 @@ abstract contract PoRBaseTest is Test {
     asset.burn(assetsHolder, amount);
   }
 
-  function _setPoRAnswer(MockERC20 asset, uint256 answer) internal {
+  function _setPoRAnswer(MockERC20 asset, int256 answer) internal {
     if (address(asset) == address(asset_1)) {
-      feed_1.setAnswer(int256(answer));
+      feed_1.setAnswer(answer);
     } else if (address(asset) == address(asset_2)) {
-      feed_2.setAnswer(int256(answer));
+      feed_2.setAnswer(answer);
     } else {
-      feed_3.setAnswer(int256(answer));
+      feed_3.setAnswer(answer);
     }
   }
